wm: Fix menu only opening once on legacy apps
To reproduce:
1. Open pre-honeycomb app with options menu
2. adb shell input keyevent MENU
3. Observe menu opening
4. adb shell input keyevent MENU
5. Observe menu closing
6. adb shell input keyevent MENU
7. Menu not re-appearing!

Root cause:
If opening menu for the first time, decorView will be null and we call
initializePanelContents to populate st.shownPanelView, therefore
hasPanelItems succeeds as it should. However, when opening menu for a
second time, decorView will exist, we will skip the code block and
shownPanelView will never be initialized, which will cause false
negative of hasPanelItems(), therefore menu will not be shown.

Fixes: 3c2bfc88 ("CP: Don't create panel if content is empty")
Test: Reproduce steps now successfully open menu every time.
Change-Id: I2c49b7595bd0f13d9800ec7f0d493ff96236adf2
Signed-off-by: Dmitrii <bankersenator@gmail.com>
Signed-off-by: Jis G Jacob <studiokeys@blissroms.org>